platform: Add nrf54l targets lc10a and lv10a
The nrf54lc10a and nrf54lv10a PSA arch test targets are identical to the
existing nrf54l15 target. Add both and move the shared configuration into
a common tgt_dev_apis_tfm_nrf54l_common directory, leaving thin per-target
wrappers (the framework discovers targets by directory and loads
target.cmake per target).
