+++ api-tests/platform/targets/tgt_dev_apis_tfm_nrf54l15/nspe/pal_storage_config.h
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2024, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2024, 2026, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,10 +15,5 @@
  * limitations under the License.
 **/
 
-#ifndef _PAL_STORAGE_CONFIG_H_
-#define _PAL_STORAGE_CONFIG_H_
-
-/* Platform specific max UID's size */
-#define ARCH_TEST_STORAGE_UID_MAX_SIZE 512
-
-#endif /* _PAL_STORAGE_CONFIG_H_ */
+/* nRF54L family shared PSA arch test configuration, see tgt_dev_apis_tfm_nrf54l_common. */
+#include "../../tgt_dev_apis_tfm_nrf54l_common/nspe/pal_storage_config.h"
